# PERF: Index.sort_values for already sorted index