@phofl 2023-11-23
Can you try this when this isn't cached? E.g. recreating the series for every pass